% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sampling_funs.R
\name{add_info}
\alias{add_info}
\title{Add sampling info to layers}
\usage{
add_info(layer, info)
}
\arguments{
\item{layer}{An 'sf' class object to add information to}

\item{info}{A data.frame with each row representing a new attribute field.
Must have columns called; 'name', 'type', 'default'. Put the name of the
attribute field in the rows of the 'name' column, the type of data in the
'type' column (text, integer, real, or logical), and the desired
default value in the 'default' column.}
}
\value{
A 'sf' object with the added information added.
}
\description{
Add attributes and fields to layers that you will
want to collect in the field. Pass in a data.frame with each row
a unique attribute field with the name spacified (e.g. comment), the type of
data (e.g. text), and the default value (e.g. foobar).
}
