% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_sampling_funs.R
\name{get_centroids}
\alias{get_centroids}
\title{Get centroids of a grid}
\usage{
get_centroids(grid_dat)
}
\arguments{
\item{grid_dat}{An 'sf' object with grid cells for which to get centroid for.}
}
\value{
An 'sf' class grid object
}
\description{
Takes a grid object and returns the centroids of each cell.
}
