% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sampling_funs.R
\name{get_data}
\alias{get_data}
\title{Import data from path and convert to UTM}
\usage{
get_data(data_path)
}
\arguments{
\item{data_path}{Path to data for import.}
}
\value{
Data projected in UTM coordinates
}
\description{
Imports data from specified path. Calculates the UTM
code for WGS84 (i.e. 32612, 32614) for the specified data and projects
the data. Returns the data projected into the correct UTM zone.
}
