% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transect_funs.R
\name{get_samp_pts}
\alias{get_samp_pts}
\title{Randomly sample points on a line}
\usage{
get_samp_pts(t_sp, n, epsg)
}
\arguments{
\item{t_sp}{Multilinestring object of class 'sf'}

\item{n}{Approximate number of sample points. See sp::spsample()}

\item{epsg}{Desired EPSG coordinate reference system code for output}
}
\value{
'sf' class object with sample points
}
\description{
Function for returning a random sample with a specified
(approximate) number of points. Must provide a line layer as a 'sf'
object class and the EPSG code to define the coordinate reference
system of the returned points.
}
