% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_sampling_funs.R
\name{make_grid}
\alias{make_grid}
\title{Makes a grid across a polygon boundary}
\usage{
make_grid(poly_layer_path, cell_size)
}
\arguments{
\item{poly_layer_path}{Character, path to the .shp file of the polygon layer
of the area of interest.}

\item{cell_size}{Numeric, the size of the grid cell in meters.}
}
\value{
An 'sf' class grid object
}
\description{
Takes a path to a polygon layer and creates a
grid with cells of a user specified dimension (in meters)
across the area. Returns an 'sf' class object.
}
