% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transect_funs.R
\name{make_transect_pts}
\alias{make_transect_pts}
\title{Make points at center of transect cells}
\usage{
make_transect_pts(t_lines, t_length, t_size)
}
\arguments{
\item{t_lines}{An 'sf' object with transect lines.}

\item{t_length}{The length of the transects in 't_layer' (in meters).}

\item{t_size}{The size of the transect cells (in meters), assumed to be
square (t_size x t_size).}
}
\value{
An 'sf' object with sampling points along transect lines.
}
\description{
Lays points out at the center of sampling cells along a transect.
User supplies a layer with transect lines, the length of each transect, and the
size (assumed to be a square) of the transect cells. Returns a point layer with
a transect ID and distance along transect to form a unique point ID.
}
