% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transect_funs.R
\name{make_wagonwheel_transects}
\alias{make_wagonwheel_transects}
\title{Make transect lines for sampling}
\usage{
make_wagonwheel_transects(origin_layer_path, t_length, spokes, t_size)
}
\arguments{
\item{origin_layer_path}{Character, path to the .shp file of the polygon or point layer to build
transects from.}

\item{t_length}{Numeric, the desired length of transects (in meters).}

\item{spokes}{Numeric, the number of spokes (number of transects) in the wheel design.}

\item{t_size}{Numeric, the size (Y x Y) of the transect cells (in meters).}
}
\value{
A 'sf' object with transect lines and ID numbers in lat long.
}
\description{
Create transects from an origin point in a wagonwheel pattern.
}
